fix: split live-room boids into its own stories file with correct meta
The previous LiveRoom story was a named export spreading a second meta
object — Storybook ignores meta on named exports, so it was rendering
BoidsSpike (the mock component) instead of BoidLiveRoom. useRawCoordStream
was never called, no socket opened, no badge rendered.

Fix: extract shared rendering into stories/boids-shared.tsx, keep mock
stories in BoidsSpike.stories.tsx, and move the live connection into
BoidsSpikeLive.stories.tsx with its own default export. Now appears as
Spikes/BoidsSpikeLive with Aloof and Clingy presets.
